+++ adapters/sp1/groth16-verifier/Cargo.toml
@@ -25,5 +25,5 @@ thiserror.workspace = true
 sp1-verifier = "5.0.0"
 hex.workspace = true
 zkaleido.workspace = true
-# serde_json = "1.0.140"
-# bincode.workspace = true
+serde_json = "1.0.140"
+bincode.workspace = true
